examples: Tier A + Tier B starter programs (38 in total)
Adds runnable example binaries under examples/ covering both the
fundamentals (Tier A: pinning, manifests, feeds, SOC/GSOC, ACT, PSS,
postage, offline helpers) and a set of small starter projects (Tier B:
swarm-paste, swarm-deploy, swarm-blog, swarm-fs, swarm-chat,
swarm-vault, swarm-share, swarm-pinner, swarm-feed-rss, swarm-relay,
swarm-keyring, swarm-cost-monitor) plus an examples/README.md that
groups everything by primitive and adds setup / caveats notes.

Also pulls in three small library fixes the example suite surfaced:

- pkg/debug/node.go: bump SupportedAPIVersion to 8.0.0 (and
  SupportedBeeVersionExact to 2.7.2-rc1-83612d37) — the Bee 2.7.x
  series moved the API major.
- pkg/debug/stake.go: RedistributionStateResponse.LastSampleDurationSeconds
  is float64; Bee returns it as a fractional float, so uint64 failed
  to deserialize.
- pkg/api/grantee.go: GetGrantees now accepts both the modern Bee
  array response and the older {"grantees": [...]} object.

Example fixes:

- examples/upload-picture: use filepath.Base() for the /bzz name —
  raw paths with slashes produce a 400.
- examples/integration-check: reuse BEE_SIGNER_HEX + a stable feed
  topic (fresh /feeds lookups can take >30s on a live network);
  retry FetchLatestFeedUpdate / IsFeedRetrievable with backoff;
  and split the PSS check into a send + a websocket-open smoke test
  rather than requiring single-node loopback delivery.

Verified with all 38 examples running end-to-end against a local Bee
2.7.2-rc1 node (no Sepolia traffic).

+++ pkg/debug/stake.go
@@ -173,7 +173,9 @@ type RedistributionStateResponse struct {
 	LastPlayedRound           uint64   `json:"lastPlayedRound"`
 	LastFrozenRound           uint64   `json:"lastFrozenRound"`
 	LastSelectedRound         uint64   `json:"lastSelectedRound"`
-	LastSampleDurationSeconds uint64   `json:"lastSampleDurationSeconds"`
+	// Bee returns this as a fractional float (e.g. 37.96302), not an
+	// integer.
+	LastSampleDurationSeconds float64  `json:"lastSampleDurationSeconds"`
 	Block                     uint64   `json:"block"`
 	Reward                    *big.Int `json:"reward"`
 	Fees                      *big.Int `json:"fees"`
@@ -181,21 +183,21 @@ type RedistributionStateResponse struct {
 }
 
 type redistributionStateJSON struct {
-	MinimumGasFunds           string `json:"minimumGasFunds"`
-	HasSufficientFunds        bool   `json:"hasSufficientFunds"`
-	IsFrozen                  bool   `json:"isFrozen"`
-	IsFullySynced             bool   `json:"isFullySynced"`
-	Phase                     string `json:"phase"`
-	Round                     uint64 `json:"round"`
-	LastWonRound              uint64 `json:"lastWonRound"`
-	LastPlayedRound           uint64 `json:"lastPlayedRound"`
-	LastFrozenRound           uint64 `json:"lastFrozenRound"`
-	LastSelectedRound         uint64 `json:"lastSelectedRound"`
-	LastSampleDurationSeconds uint64 `json:"lastSampleDurationSeconds"`
-	Block                     uint64 `json:"block"`
-	Reward                    string `json:"reward"`
-	Fees                      string `json:"fees"`
-	IsHealthy                 bool   `json:"isHealthy"`
+	MinimumGasFunds           string  `json:"minimumGasFunds"`
+	HasSufficientFunds        bool    `json:"hasSufficientFunds"`
+	IsFrozen                  bool    `json:"isFrozen"`
+	IsFullySynced             bool    `json:"isFullySynced"`
+	Phase                     string  `json:"phase"`
+	Round                     uint64  `json:"round"`
+	LastWonRound              uint64  `json:"lastWonRound"`
+	LastPlayedRound           uint64  `json:"lastPlayedRound"`
+	LastFrozenRound           uint64  `json:"lastFrozenRound"`
+	LastSelectedRound         uint64  `json:"lastSelectedRound"`
+	LastSampleDurationSeconds float64 `json:"lastSampleDurationSeconds"`
+	Block                     uint64  `json:"block"`
+	Reward                    string  `json:"reward"`
+	Fees                      string  `json:"fees"`
+	IsHealthy                 bool    `json:"isHealthy"`
 }
 
 func (r *RedistributionStateResponse) UnmarshalJSON(b []byte) error {
